v 1.2 done.

+++ lib/features/order/data/models/order/pharmaceutical.details.dart
@@ -10,7 +10,7 @@ class Pharmaceutical extends Equatable {
   final String? manufactureCompany;
   final int? quantityAvailable;
   final String? expireDate;
-  final int? price;
+  final double? price;
   final DateTime? createdAt;
   final DateTime? updatedAt;
   final Pivot? pivot;
@@ -38,7 +38,7 @@ class Pharmaceutical extends Equatable {
       manufactureCompany: json['manufacture_company'] as String?,
       quantityAvailable: json['quantity_available'] as int?,
       expireDate: json['expire_date'] as String?,
-      price: json['price'] as int?,
+      price: json['price'] as double?,
       createdAt: json['created_at'] == null
           ? null
           : DateTime.parse(json['created_at'] as String),
